refactor(gateway): extract tools tests to tools_tests.rs to stay under 1500-line limit
tools.rs (1543 lines) exceeded the 1500-line production Rust file limit.
Moved the ~210-line #[cfg(test)] mod tests block into a separate
gateway/tools_tests.rs, reducing tools.rs to 1331 lines.

- Declared #[cfg(test)] mod tools_tests; in gateway/mod.rs
- Promoted describe_needs_refresh to pub(crate) for test access
- All 5 tests pass: cargo test -p dcc-mcp-gateway -- tools_tests
